Also move GPDB6 greenplum files into greenplum directory.
