+++ contrib/pg_upgrade/Makefile
@@ -15,9 +15,9 @@ OBJS += greenplum/aotable.o greenplum/gpdb4_heap_convert.o greenplum/version_gp.
 		greenplum/check_gp.o greenplum/file_gp.o greenplum/reporting.o \
 		greenplum/aomd_filehandler.o greenplum/option_gp.o \
 		greenplum/controldata_gp.o \
-		old_tablespace_file_contents.o old_tablespace_file_parser.o \
-		tablespace_gp.o old_tablespace_file_parser_observer.o \
-		info_gp.o
+		greenplum/old_tablespace_file_contents.o greenplum/old_tablespace_file_parser.o \
+		greenplum/tablespace_gp.o greenplum/old_tablespace_file_parser_observer.o \
+		greenplum/info_gp.o
 
 PG_CPPFLAGS  = -DFRONTEND -DDLSUFFIX=\"$(DLSUFFIX)\" -I$(srcdir) -I$(libpq_srcdir)
 PG_LIBS = $(libpq_pgport)
