+++ contrib/pg_upgrade/greenplum/old_tablespace_file_contents.c
