+++ contrib/pg_upgrade/greenplum/pg_upgrade_greenplum.h
@@ -119,4 +119,17 @@ void report_progress(ClusterInfo *cluster, progress_type op, char *fmt,...)
 pg_attribute_printf(3, 4);
 void close_progress(void);
 
+/* tablespace_gp.c */
+void populate_old_cluster_with_old_tablespaces(ClusterInfo *oldCluster, const char *file_path);
+void generate_old_tablespaces_file(ClusterInfo *oldCluster);
+void populate_gpdb6_cluster_tablespace_suffix(ClusterInfo *cluster);
+bool is_gpdb_version_with_filespaces(ClusterInfo *cluster);
+
+
+static inline bool
+is_gpdb6(ClusterInfo *cluster)
+{
+	return GET_MAJOR_VERSION(cluster->major_version) == 904;
+}
+
 #endif /* PG_UPGRADE_GREENPLUM_H */
