+++ contrib/pg_upgrade/test/unit/Makefile
@@ -2,6 +2,7 @@ subdir=contrib/pg_upgrade/test/unit
 top_srcdir=../../../..
 top_builddir=../../../..
 pg_upgrade_directory=../..
+pg_upgrade_greenplum_directory=../../greenplum
 
 OBJS = tablespace_test.o \
 	old_tablespace_file_contents_test.o \
@@ -33,16 +34,16 @@ compile_test = $(CC) $(CFLAGS) $(debugging_flags) $^ $(libpq_pgport) $(LDFLAGS) 
 tablespace_test.t: tablespace_test.o $(pg_upgrade_directory)/tablespace.o $(test_dependencies)
 	$(compile_test)
 
-old_tablespace_file_contents_test.t: old_tablespace_file_contents_test.o $(pg_upgrade_directory)/old_tablespace_file_contents.o $(test_dependencies)
+old_tablespace_file_contents_test.t: old_tablespace_file_contents_test.o $(pg_upgrade_greenplum_directory)/old_tablespace_file_contents.o $(test_dependencies)
 	$(compile_test)
 
-old_tablespace_file_parser_test.t: old_tablespace_file_parser_test.o $(pg_upgrade_directory)/old_tablespace_file_parser.o $(test_dependencies)
+old_tablespace_file_parser_test.t: old_tablespace_file_parser_test.o $(pg_upgrade_greenplum_directory)/old_tablespace_file_parser.o $(test_dependencies)
 	$(compile_test)
 
-tablespace_gp_test.t: tablespace_gp_test.o $(pg_upgrade_directory)/tablespace_gp.o $(test_dependencies)
+tablespace_gp_test.t: tablespace_gp_test.o $(pg_upgrade_greenplum_directory)/tablespace_gp.o $(test_dependencies)
 	$(compile_test)
 
-info_gp_test.t: info_gp_test.o $(pg_upgrade_directory)/info_gp.o $(test_dependencies)
+info_gp_test.t: info_gp_test.o $(pg_upgrade_greenplum_directory)/info_gp.o $(test_dependencies)
 	$(compile_test)
 
 check_gp_test.t: mock_util.o \
@@ -50,7 +51,8 @@ check_gp_test.t: mock_util.o \
 	check_gp_test.o \
 	$(MOCK_DIR)/backend/libpq/fe-exec_mock.o \
 	$(test_dependencies) \
-	$(pg_upgrade_directory)/check_gp.o
+	$(pg_upgrade_greenplum_directory)/check_gp.o \
+	$(pg_upgrade_greenplum_directory)/option_gp.o
 	$(compile_test)
 
 clean:
